feat: modularized form code into sections for improved DX
