--- app/apply/layouts/events.tsx
@@ -1,7 +1,0 @@
-export function Events() {
-    return (
-        <div>
-            <h1>Events</h1>
-        </div>
-    )
-}
