+++ app/apply/layouts/form_sections/engineering.tsx
@@ -1,0 +1,7 @@
+export function Engineering({ control }: { control: any }) {
+    return (
+        <div>
+            Engineering
+        </div>
+    )
+}
