+++ app/apply/layouts/form_sections/events.tsx
@@ -1,0 +1,7 @@
+export function Events({ control }: { control: any }) {
+    return (
+        <div>
+            Events
+        </div>
+    )
+}
