+++ app/apply/layouts/form_sections/marketing.tsx
@@ -1,0 +1,7 @@
+export function Marketing({ control }: { control: any }) {
+    return (
+        <div>
+            Marketing
+        </div>
+    )
+}
