+++ app/apply/layouts/form_sections/outreach.tsx
@@ -1,0 +1,7 @@
+export function Outreach({ control }: { control: any }) {
+    return (
+        <div>
+            Outreach
+        </div>
+    )
+}
