+++ app/apply/layouts/form_sections/podcast.tsx
@@ -1,0 +1,7 @@
+export function Podcast({ control }: { control: any }) {
+    return (
+        <div>
+            Podcast
+        </div>
+    )
+}
