+++ app/apply/layouts/form_sections/secretary.tsx
@@ -1,0 +1,7 @@
+export function Secretary({ control }: { control: any }) {
+    return (
+        <div>
+            Secretary
+        </div>
+    )
+}
